LayoutTests/http/tests/security/xssAuditor/script-tag-inside-svg-tag2-expected.txt

CONSOLE MESSAGE: Refused to execute a JavaScript script. Source code of script found within request.

 Ensures HTML entities are recognized in script blocks in a context where CDATA is allowed even with <!-- comments -->.
